+++ egui/src/containers/collapsing_header.rs
@@ -223,8 +223,11 @@ impl CollapsingHeader {
         let text_pos = available.min + vec2(ui.spacing().indent, 0.0);
         let galley = label.layout_width(ui, available.right() - text_pos.x);
         let text_max_x = text_pos.x + galley.size.x;
-        let desired_width = text_max_x + button_padding.x - available.left();
-        let desired_width = desired_width.max(available.width());
+
+        let mut desired_width = text_max_x + button_padding.x - available.left();
+        if ui.visuals().collapsing_header_frame {
+            desired_width = desired_width.max(available.width()); // fill full width
+        }
 
         let mut desired_size = vec2(desired_width, galley.size.y + 2.0 * button_padding.y);
         desired_size = desired_size.at_least(ui.spacing().interact_size);
@@ -246,13 +249,16 @@ impl CollapsingHeader {
 
         let visuals = ui.style().interact(&header_response);
         let text_color = visuals.text_color();
-        ui.painter().add(Shape::Rect {
-            rect: header_response.rect.expand(visuals.expansion),
-            corner_radius: visuals.corner_radius,
-            fill: visuals.bg_fill,
-            stroke: visuals.bg_stroke,
-            // stroke: Default::default(),
-        });
+
+        if ui.visuals().collapsing_header_frame {
+            ui.painter().add(Shape::Rect {
+                rect: header_response.rect.expand(visuals.expansion),
+                corner_radius: visuals.corner_radius,
+                fill: visuals.bg_fill,
+                stroke: visuals.bg_stroke,
+                // stroke: Default::default(),
+            });
+        }
 
         {
             let (mut icon_rect, _) = ui.spacing().icon_rectangles(header_response.rect);