+++ egui/src/ui.rs
@@ -137,6 +137,11 @@ impl Ui {
         self.style = style.into();
     }
 
+    /// Reset to the default style set in [`Context`].
+    pub fn reset_style(&mut self) {
+        self.style = self.ctx().style();
+    }
+
     /// The current spacing options for this `Ui`.
     /// Short for `ui.style().spacing`.
     #[inline(always)]
@@ -215,11 +220,41 @@ impl Ui {
     /// ```
     pub fn set_enabled(&mut self, enabled: bool) {
         self.enabled &= enabled;
-        if self.enabled {
-            self.painter.set_fade_to_color(None);
-        } else {
+        if !self.enabled && self.visible() {
             self.painter
                 .set_fade_to_color(Some(self.visuals().window_fill()));
+        }
+    }
+
+    /// If `false`, any widgets added to the `Ui` will be invisible and non-interactive.
+    #[inline(always)]
+    pub fn visible(&self) -> bool {
+        self.painter.visible()
+    }
+
+    /// Calling `set_visible(false)` will cause all further widgets to be invisible,
+    /// yet still allocate space.
+    ///
+    /// The widgets will not be interactive (`set_visible(false)` implies `set_enabled(false)`).
+    ///
+    /// Calling `set_visible(true)` has no effect.
+    ///
+    /// ### Example
+    /// ```
+    /// # let ui = &mut egui::Ui::__test();
+    /// # let mut visible = true;
+    /// ui.group(|ui|{
+    ///     ui.checkbox(&mut visible, "Show subsection");
+    ///     ui.set_visible(visible);
+    ///     if ui.button("Button that is not always shown").clicked() {
+    ///         /* … */
+    ///     }
+    /// });
+    /// ```
+    pub fn set_visible(&mut self, visible: bool) {
+        self.set_enabled(visible);
+        if !visible {
+            self.painter.set_invisible();
         }
     }
 
@@ -663,9 +698,9 @@ impl Ui {
         widget_rect
     }
 
-    /// Allocate a specific part of the `Ui‘.
+    /// Allocate a specific part of the `Ui`.
     ///
-    /// Ignore the layout of the `Ui‘: just put my widget here!
+    /// Ignore the layout of the `Ui`: just put my widget here!
     /// The layout cursor will advance to past this `rect`.
     pub fn allocate_rect(&mut self, rect: Rect, sense: Sense) -> Response {
         let id = self.advance_cursor_after_rect(rect);
@@ -1226,6 +1261,8 @@ impl Ui {
     ///     ui.label("Within a frame");
     /// });
     /// ```
+    ///
+    /// Se also [`Self::scope`].
     pub fn group<R>(&mut self, add_contents: impl FnOnce(&mut Ui) -> R) -> InnerResponse<R> {
         crate::Frame::group(self.style()).show(self, add_contents)
     }
@@ -1243,7 +1280,9 @@ impl Ui {
     /// ```
     pub fn scope<R>(&mut self, add_contents: impl FnOnce(&mut Ui) -> R) -> InnerResponse<R> {
         let child_rect = self.available_rect_before_wrap();
+        let next_auto_id_source = self.next_auto_id_source;
         let mut child_ui = self.child_ui(child_rect, *self.layout());
+        self.next_auto_id_source = next_auto_id_source; // HACK: we want `scope` to only increment this once, so that `ui.scope` is equivalent to `ui.allocate_space`.
         let ret = add_contents(&mut child_ui);
         let response = self.allocate_rect(child_ui.min_rect(), Sense::hover());
         InnerResponse::new(ret, response)
@@ -1285,6 +1324,9 @@ impl Ui {
     }
 
     /// Create a child ui which is indented to the right.
+    ///
+    /// The `id_source` here be anything at all.
+    // TODO: remove `id_source` argument?
     #[inline(always)]
     pub fn indent<R>(
         &mut self,
@@ -1315,7 +1357,8 @@ impl Ui {
         };
         let ret = add_contents(&mut child_ui);
 
-        let end_with_horizontal_line = true;
+        let end_with_horizontal_line = self.spacing().indent_ends_with_horizontal_line;
+
         if end_with_horizontal_line {
             child_ui.add_space(4.0);
         }