+++ egui_demo_lib/src/apps/demo/plot_demo.rs
@@ -1,5 +1,8 @@
-use egui::plot::{Line, MarkerShape, Plot, Points, Value, Values};
 use egui::*;
+use plot::{
+    Arrows, Corner, HLine, Legend, Line, MarkerShape, Plot, PlotImage, Points, Polygon, Text,
+    VLine, Value, Values,
+};
 use std::f64::consts::TAU;
 
 #[derive(PartialEq)]
@@ -9,19 +12,17 @@ struct LineDemo {
     circle_radius: f64,
     circle_center: Pos2,
     square: bool,
-    legend: bool,
     proportional: bool,
 }
 
 impl Default for LineDemo {
     fn default() -> Self {
         Self {
-            animate: true,
+            animate: !cfg!(debug_assertions),
             time: 0.0,
             circle_radius: 1.5,
             circle_center: Pos2::new(0.0, 0.0),
             square: false,
-            legend: true,
             proportional: true,
         }
     }
@@ -35,7 +36,6 @@ impl LineDemo {
             circle_radius,
             circle_center,
             square,
-            legend,
             proportional,
             ..
         } = self;
@@ -47,7 +47,7 @@ impl LineDemo {
                     ui.add(
                         egui::DragValue::new(circle_radius)
                             .speed(0.1)
-                            .clamp_range(0.0..=f32::INFINITY)
+                            .clamp_range(0.0..=f64::INFINITY)
                             .prefix("r: "),
                     );
                     ui.horizontal(|ui| {
@@ -68,9 +68,10 @@ impl LineDemo {
             ui.vertical(|ui| {
                 ui.style_mut().wrap = Some(false);
                 ui.checkbox(animate, "animate");
-                ui.checkbox(square, "square view");
-                ui.checkbox(legend, "legend");
-                ui.checkbox(proportional, "proportional data axes");
+                ui.checkbox(square, "square view")
+                    .on_hover_text("Always keep the viewport square.");
+                ui.checkbox(proportional, "Proportional data axes")
+                    .on_hover_text("Tick are the same size on both axes.");
             });
         });
     }
@@ -94,7 +95,7 @@ impl LineDemo {
         let time = self.time;
         Line::new(Values::from_explicit_callback(
             move |x| 0.5 * (2.0 * x).sin() * time.sin(),
-            f64::NEG_INFINITY..=f64::INFINITY,
+            ..,
             512,
         ))
         .color(Color32::from_rgb(200, 100, 100))
@@ -120,11 +121,11 @@ impl Widget for &mut LineDemo {
             ui.ctx().request_repaint();
             self.time += ui.input().unstable_dt.at_most(1.0 / 30.0) as f64;
         };
-        let mut plot = Plot::new("Lines Demo")
+        let mut plot = Plot::new("lines_demo")
             .line(self.circle())
             .line(self.sin())
             .line(self.thingy())
-            .show_legend(self.legend);
+            .legend(Legend::default());
         if self.square {
             plot = plot.view_aspect(1.0);
         }
@@ -139,7 +140,7 @@ impl Widget for &mut LineDemo {
 struct MarkerDemo {
     fill_markers: bool,
     marker_radius: f32,
-    custom_marker_color: bool,
+    automatic_colors: bool,
     marker_color: Color32,
 }
 
@@ -148,8 +149,8 @@ impl Default for MarkerDemo {
         Self {
             fill_markers: true,
             marker_radius: 5.0,
-            custom_marker_color: false,
-            marker_color: Color32::GRAY,
+            automatic_colors: true,
+            marker_color: Color32::GREEN,
         }
     }
 }
@@ -174,7 +175,7 @@ impl MarkerDemo {
                 .radius(self.marker_radius)
                 .shape(marker);
 
-                if self.custom_marker_color {
+                if !self.automatic_colors {
                     points = points.color(self.marker_color);
                 }
 
@@ -187,20 +188,22 @@ impl MarkerDemo {
 impl Widget for &mut MarkerDemo {
     fn ui(self, ui: &mut Ui) -> Response {
         ui.horizontal(|ui| {
-            ui.checkbox(&mut self.fill_markers, "fill markers");
+            ui.checkbox(&mut self.fill_markers, "Fill");
             ui.add(
                 egui::DragValue::new(&mut self.marker_radius)
                     .speed(0.1)
-                    .clamp_range(0.0..=f32::INFINITY)
-                    .prefix("marker radius: "),
+                    .clamp_range(0.0..=f64::INFINITY)
+                    .prefix("Radius: "),
             );
-            ui.checkbox(&mut self.custom_marker_color, "custom marker color");
-            if self.custom_marker_color {
+            ui.checkbox(&mut self.automatic_colors, "Automatic colors");
+            if !self.automatic_colors {
                 ui.color_edit_button_srgba(&mut self.marker_color);
             }
         });
 
-        let mut markers_plot = Plot::new("Markers Demo").data_aspect(1.0);
+        let mut markers_plot = Plot::new("markers_demo")
+            .data_aspect(1.0)
+            .legend(Legend::default());
         for marker in self.markers() {
             markers_plot = markers_plot.points(marker);
         }
@@ -208,10 +211,148 @@ impl Widget for &mut MarkerDemo {
     }
 }
 
+#[derive(PartialEq)]
+struct LegendDemo {
+    config: Legend,
+}
+
+impl Default for LegendDemo {
+    fn default() -> Self {
+        Self {
+            config: Legend::default(),
+        }
+    }
+}
+
+impl LegendDemo {
+    fn line_with_slope(slope: f64) -> Line {
+        Line::new(Values::from_explicit_callback(move |x| slope * x, .., 100))
+    }
+    fn sin() -> Line {
+        Line::new(Values::from_explicit_callback(move |x| x.sin(), .., 100))
+    }
+    fn cos() -> Line {
+        Line::new(Values::from_explicit_callback(move |x| x.cos(), .., 100))
+    }
+}
+
+impl Widget for &mut LegendDemo {
+    fn ui(self, ui: &mut Ui) -> Response {
+        let LegendDemo { config } = self;
+
+        egui::Grid::new("settings").show(ui, |ui| {
+            ui.label("Text style:");
+            ui.horizontal(|ui| {
+                TextStyle::all().for_each(|style| {
+                    ui.selectable_value(&mut config.text_style, style, format!("{:?}", style));
+                });
+            });
+            ui.end_row();
+
+            ui.label("Position:");
+            ui.horizontal(|ui| {
+                Corner::all().for_each(|position| {
+                    ui.selectable_value(&mut config.position, position, format!("{:?}", position));
+                });
+            });
+            ui.end_row();
+
+            ui.label("Opacity:");
+            ui.add(
+                egui::DragValue::new(&mut config.background_alpha)
+                    .speed(0.02)
+                    .clamp_range(0.0..=1.0),
+            );
+            ui.end_row();
+        });
+
+        let legend_plot = Plot::new("legend_demo")
+            .line(LegendDemo::line_with_slope(0.5).name("lines"))
+            .line(LegendDemo::line_with_slope(1.0).name("lines"))
+            .line(LegendDemo::line_with_slope(2.0).name("lines"))
+            .line(LegendDemo::sin().name("sin(x)"))
+            .line(LegendDemo::cos().name("cos(x)"))
+            .legend(*config)
+            .data_aspect(1.0);
+        ui.add(legend_plot)
+    }
+}
+
+#[derive(PartialEq, Default)]
+struct ItemsDemo {}
+
+impl ItemsDemo {}
+
+impl Widget for &mut ItemsDemo {
+    fn ui(self, ui: &mut Ui) -> Response {
+        let n = 100;
+        let mut sin_values: Vec<_> = (0..=n)
+            .map(|i| remap(i as f64, 0.0..=n as f64, -TAU..=TAU))
+            .map(|i| Value::new(i, i.sin()))
+            .collect();
+
+        let line = Line::new(Values::from_values(sin_values.split_off(n / 2))).fill(-1.5);
+        let polygon = Polygon::new(Values::from_parametric_callback(
+            |t| (4.0 * t.sin() + 2.0 * t.cos(), 4.0 * t.cos() + 2.0 * t.sin()),
+            0.0..TAU,
+            100,
+        ));
+        let points = Points::new(Values::from_values(sin_values))
+            .stems(-1.5)
+            .radius(1.0);
+
+        let arrows = {
+            let pos_radius = 8.0;
+            let tip_radius = 7.0;
+            let arrow_origins = Values::from_parametric_callback(
+                |t| (pos_radius * t.sin(), pos_radius * t.cos()),
+                0.0..TAU,
+                36,
+            );
+            let arrow_tips = Values::from_parametric_callback(
+                |t| (tip_radius * t.sin(), tip_radius * t.cos()),
+                0.0..TAU,
+                36,
+            );
+            Arrows::new(arrow_origins, arrow_tips)
+        };
+        let image = PlotImage::new(
+            TextureId::Egui,
+            Value::new(0.0, 10.0),
+            [
+                ui.fonts().texture().width as f32 / 100.0,
+                ui.fonts().texture().height as f32 / 100.0,
+            ],
+        );
+
+        let plot = Plot::new("items_demo")
+            .hline(HLine::new(9.0).name("Lines horizontal"))
+            .hline(HLine::new(-9.0).name("Lines horizontal"))
+            .vline(VLine::new(9.0).name("Lines vertical"))
+            .vline(VLine::new(-9.0).name("Lines vertical"))
+            .line(line.name("Line with fill"))
+            .polygon(polygon.name("Convex polygon"))
+            .points(points.name("Points with stems"))
+            .text(Text::new(Value::new(-3.0, -3.0), "wow").name("Text"))
+            .text(Text::new(Value::new(-2.0, 2.5), "so graph").name("Text"))
+            .text(Text::new(Value::new(3.0, 3.0), "much color").name("Text"))
+            .text(Text::new(Value::new(2.5, -2.0), "such plot").name("Text"))
+            .image(image.name("Image"))
+            .arrows(arrows.name("Arrows"))
+            .legend(Legend::default().position(Corner::RightBottom))
+            .show_x(false)
+            .show_y(false)
+            .data_aspect(1.0);
+        ui.add(plot)
+    }
+}
+
 #[derive(PartialEq, Eq)]
 enum Panel {
     Lines,
     Markers,
+    Legend,
+    Items,
 }
 
 impl Default for Panel {
@@ -224,6 +365,8 @@ impl Default for Panel {
 pub struct PlotDemo {
     line_demo: LineDemo,
     marker_demo: MarkerDemo,
+    legend_demo: LegendDemo,
+    items_demo: ItemsDemo,
     open_panel: Panel,
 }
 
@@ -244,23 +387,27 @@ impl super::Demo for PlotDemo {
 
 impl super::View for PlotDemo {
     fn ui(&mut self, ui: &mut Ui) {
-        ui.vertical_centered(|ui| {
+        ui.horizontal(|ui| {
             egui::reset_button(ui, self);
-            ui.add(crate::__egui_github_link_file!());
-            ui.label("Pan by dragging, or scroll (+ shift = horizontal).");
-            if cfg!(target_arch = "wasm32") {
-                ui.label("Zoom with ctrl / ⌘ + mouse wheel, or with pinch gesture.");
-            } else if cfg!(target_os = "macos") {
-                ui.label("Zoom with ctrl / ⌘ + scroll.");
-            } else {
-                ui.label("Zoom with ctrl + scroll.");
-            }
-            ui.label("Reset view with double-click.");
+            ui.collapsing("Instructions", |ui| {
+                ui.label("Pan by dragging, or scroll (+ shift = horizontal).");
+                if cfg!(target_arch = "wasm32") {
+                    ui.label("Zoom with ctrl / ⌘ + mouse wheel, or with pinch gesture.");
+                } else if cfg!(target_os = "macos") {
+                    ui.label("Zoom with ctrl / ⌘ + scroll.");
+                } else {
+                    ui.label("Zoom with ctrl + scroll.");
+                }
+                ui.label("Reset view with double-click.");
+                ui.add(crate::__egui_github_link_file!());
+            });
         });
         ui.separator();
         ui.horizontal(|ui| {
             ui.selectable_value(&mut self.open_panel, Panel::Lines, "Lines");
             ui.selectable_value(&mut self.open_panel, Panel::Markers, "Markers");
+            ui.selectable_value(&mut self.open_panel, Panel::Legend, "Legend");
+            ui.selectable_value(&mut self.open_panel, Panel::Items, "Items");
         });
         ui.separator();
 
@@ -270,6 +417,12 @@ impl super::View for PlotDemo {
             }
             Panel::Markers => {
                 ui.add(&mut self.marker_demo);
+            }
+            Panel::Legend => {
+                ui.add(&mut self.legend_demo);
+            }
+            Panel::Items => {
+                ui.add(&mut self.items_demo);
             }
         }
     }