+++ eframe/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to the `eframe` crate.
 ## Unreleased
 
 
+## 0.13.1 - 2021-06-24
+
+* Fix `http` feature flag and docs
+
+
+## 0.13.0 - 2021-06-24
+
+* `App::setup` now takes a `Frame` and `Storage` by argument.
+* `App::load` has been removed. Implement `App::setup` instead.
+
+
 ## 0.12.0 - 2021-05-10
 * Moved options out of `trait App` into new `NativeOptions`.
 * Add option for `always_on_top`.